plot a BN as a picture.

plot a BN as a picture.

# S3 method for BN
plot(
  x,
  method = "default",
  use.node.names = TRUE,
  frac = 0.2,
  max.weight = max(dag(x)),
  node.size.lab = 14,
  node.col = rep("white", num.nodes(x)),
  plot.wpdag = FALSE,
  ...
)

Arguments

x

a BN object.

method

either default of qgraph. The default method requires the Rgraphviz package, while qgraph requires the qgraph package and allows for a greater customization.

use.node.names

TRUE if node names have to be printed. If FALSE, numbers are used instead.

frac

minimum fraction [0,1] of presence of an edge to be plotted (used in case of plot.wpdag=TRUE).

max.weight

maximum possible weight of an edge (used in case of plot.wpdag=TRUE).

node.size.lab

font size for the node labels in the default mode.

node.col

list of (R) colors for the nodes.

plot.wpdag

if TRUE plot the network according to the WPDAG computed using bootstrap instead of the DAG.

...

potential further arguments when using method="qgraph". Please refer to the qgraph documentation for the parameters available for the qgraph() method.